Switch to consistent use of goal object in data contexts
Makes it easy to use the same template helpers in goal tree, goal
card, subgoal row, and subgoal menu. Also more semantic, e.g. in
subgoal_row extra properties of the menu do not get mixed with the
"goal" context.

Causes some problems with parent contexts in each blocks; these
introduce an extra data context set to the goal object. But in general
this approach is more semantic and granular.